requirements.txt.txt
streamlit
pytesseract
opencv-python
pillow
dateparser
apscheduler
python-dotenv
easyocr